% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaPlotBars.R
\name{XiaPlotBars}
\alias{XiaPlotBars}
\title{XiaPlotBars}
\usage{
XiaPlotBars(W, out, main = "Composition Fitting Graph", niter = 1000,
  diagV = FALSE, smallV = FALSE, sub = FALSE)
}
\arguments{
\item{W}{raw count matrix used in model fit}

\item{out}{model fit from LNM.EM}

\item{main}{desired title for plot, defaults to Composition Fitting graph}

\item{niter}{number of simulations for bars, defaults to 1000}

\item{diagV}{boolean of whether to use diagonal Sigma for simulations, defaults to FALSE}

\item{smallV}{boolean of whether to use diagonal Sigma with very small entries for simulations, defaults to false}

\item{sub}{desired subtitle for plot, defaults to MSPE}
}
\description{
Function to plot Xia model fit, with empirical 95 percent bars
}
