% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaPlotWBars.R
\name{XiaPlotWBars}
\alias{XiaPlotWBars}
\title{XiaPlotWBars}
\usage{
XiaPlotWBars(W, out, main = "Composition Fitting Graph", niter = 1000)
}
\arguments{
\item{W}{raw count matrix used in model fit}

\item{out}{model fit from LNM.EM}

\item{main}{desired title for plot, defaults to Composition Fitting Graph}

\item{niter}{number of simulations for bars, defaults to 1000}
}
\description{
Function to plot Xia model fit for counts W, with empirical 95 percent points
}
