% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsim.R
\name{Xsim}
\alias{Xsim}
\title{Xsim}
\usage{
Xsim(out, W, niter = 1000)
}
\arguments{
\item{out}{model fit from LNM.EM}

\item{W}{raw count matrix used in model fit}

\item{niter}{number of simulations, defaults to 1000}
}
\description{
Function to simulate raw compositions X from model fit
}
